% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{markovreport}
\alias{markovreport}
\title{Produce attribution report}
\usage{
markovreport(data, date_from, date_to, look_back, channels, goal_number)
}
\arguments{
\item{data}{The name of the data frame to use .}

\item{date_from}{The first full day to report, in the format "YYYY-MM-DD" .}

\item{date_to}{The last full day to report, in the format "YYYY-MM-DD" .}

\item{look_back}{The lookback window, in days .}

\item{channels}{The name of the variable that contains the cleaned source .}

\item{goal_number}{The number of the Google Analytics conversion that is a purchase .}
}
\description{
This is a wrapper to channelAttribution that takes a clickstream dataframe and produces a .
}
\examples{
markovreport(data, date_from, date_to, look_back, channels, goal_number)
markovreport(mydataframe, "2019-06-01", "2019-06-30", 30, "my_source_variable", 4)
}
\keyword{attribution,}
\keyword{clickstream}
\keyword{markov,}
